+++ src/main/java/pl/pmar/blogplatform/service/UserService.java
@@ -1,45 +1,57 @@
 package pl.pmar.blogplatform.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pl.pmar.blogplatform.model.Post;
+import pl.pmar.blogplatform.repository.PostRepository;
 import pl.pmar.blogplatform.repository.UserRepository;
 import pl.pmar.blogplatform.model.User;
-
-import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
 
     private final UserRepository userRepository;
-    private final PostService postService;
+    private final PostRepository postRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository, PostService postService) {
+    public UserService(UserRepository userRepository, PostRepository postRepository) {
         this.userRepository = userRepository;
-        this.postService = postService;
+        this.postRepository = postRepository;
     }
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public ResponseEntity<List<User>> getAllUsers() {
+        List<User> users = userRepository.findAll();
+        return ResponseEntity.ok(users);
     }
 
-    public User getUserById(Integer id) {
-        return userRepository.findById(id).orElse(null);
+    public ResponseEntity<User> getUserById(Integer id) {
+        Optional<User> user = userRepository.findById(id);
+        return user
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
 
     }
 
-    public User saveUser(User user) {
-        return userRepository.save(user);
+    public ResponseEntity<User> saveUser(User user) {
+        User savedUser = userRepository.save(user);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedUser);
     }
 
-    public void deleteUser(Integer id) {
-        List<Post> userPosts = postService.getUserPosts(id);
-        for (Post post : userPosts) {
-            postService.deletePost(post.getId());
+    public ResponseEntity<Void> deleteUser(Integer id) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        } else {
+            List<Post> userPosts = postRepository.findAllByUserId(id);
+            postRepository.deleteAll(userPosts);
+            userRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
         }
-        userRepository.deleteById(id);
+
     }
 
 }
